automatically apply OneVsAll in CSvm (renamed from CSvmBinaryClassifierBuilder)
