+++ NeoML/docs/en/API/ClassificationAndRegression/README.md
@@ -37,7 +37,7 @@ It is implemented by the [CLinearBinaryClassifierBuilder](Linear.md) class that 
 
 Support-vector machine translates the input data into vectors in a high-dimensional space and searches for a maximum-margin dividing hyperplane.
 
-It is implemented by the [CSvmBinaryClassifierBuilder](Svm.md) class. The trained model is represented by the `ILinearBinaryModel` or `ISvmBinaryModel` interface, depending on the type of kernel used.
+It is implemented by the [CSvm](Svm.md) class. The trained model is represented by the [`ILinearBinaryModel`](Linear.md#for-classification), `ISvmBinaryModel` or [`IOneVersusAllModel`](OneVersusAll.md#model) interface, depending on the type of kernel used and number of classes.
 
 ### Decision tree
