+++ NeoML/docs/ru/API/ClassificationAndRegression/Svm.md
@@ -1,17 +1,17 @@
-# Машина опорных векторов CSvmBinaryClassifierBuilder
+# Машина опорных векторов CSvm
 
-- [Машина опорных векторов CSvmBinaryClassifierBuilder](#машина-опорных-векторов-csvmbinaryclassifierbuilder)
+- [Машина опорных векторов CSvm](#машина-опорных-векторов-csvm)
 	- [Параметры построения модели](#параметры-построения-модели)
 	- [Модель](#модель)
 	- [Пример](#пример)
 
 Метод опорных векторов переводит исходные вектора в пространство более высокой размерности и ищет в нем разделяющую гиперплоскость с максимальным зазором.
 
-В **NeoML** алгоритм реализован классом `CSvmBinaryClassifierBuilder`. Он предоставляет метод `Train` для обучения модели бинарной классификации.
+В **NeoML** алгоритм реализован классом `CSvm`. Он предоставляет метод `Train` для обучения модели бинарной классификации.
 
 ## Параметры построения модели
 
-Параметры реализованы структурой `CSvmBinaryClassifierBuilder::CParams`.
+Параметры реализованы структурой `CSvm::CParams`.
 
 - *KernelType* — тип используемого ядра;
 - *ErrorWeight* — вес "ошибки" относительно регуляризатора;
@@ -24,7 +24,7 @@
 
 ## Модель
 
-Построенная модель может описываться интерфейсом [`ILinearBinaryModel`](Linear.md#для-классификации) при использовании ядра `KT_Linear` и интерфейсом `ISvmBinaryModel` в остальных случаях.
+Построенная модель может описываться интерфейсом [`ILinearBinaryModel`](Linear.md#для-классификации) при использовании ядра `KT_Linear`, [`IOneVersusAllModel`](OneVersusAll.md#model) если количество классов в обучающей выборке превышает 2; и интерфейсом `ISvmBinaryModel` в остальных случаях.
 
 ```c++
 // интерфейс SVM binary классификатора
@@ -56,8 +56,8 @@ public:
 ```c++
 CPtr<Model> buildModel( IProblem* data )
 {
-	CSvmBinaryClassifierBuilder::CParams params( CSvmKernel::KT_RBF );
-	CSvmBinaryClassifierBuilder builder( params );
+	CSvm::CParams params( CSvmKernel::KT_RBF );
+	CSvm builder( params );
 	return builder.Train( *data );
 }
 ```
