+++ NeoML/src/CMakeLists.txt
@@ -150,7 +150,7 @@ target_sources( ${PROJECT_NAME} PRIVATE
     TraditionalML/SparseFloatMatrix.cpp
     TraditionalML/SparseFloatVector.cpp
     TraditionalML/StratifiedCrossValidationSubProblem.cpp
-    TraditionalML/SvmBinaryClassifierBuilder.cpp
+    TraditionalML/Svm.cpp
     TraditionalML/SvmBinaryModel.cpp
     TraditionalML/SvmBinaryModel.h
     TraditionalML/SvmKernel.cpp
@@ -268,6 +268,7 @@ target_sources( ${PROJECT_NAME} PRIVATE
     ../include/NeoML/TraditionalML/SparseFloatMatrix.h
     ../include/NeoML/TraditionalML/SparseFloatVector.h
     ../include/NeoML/TraditionalML/SparseVectorIterator.h
+    ../include/NeoML/TraditionalML/Svm.h
     ../include/NeoML/TraditionalML/SvmBinaryClassifierBuilder.h
     ../include/NeoML/TraditionalML/SvmKernel.h
     ../include/NeoML/TraditionalML/StratifiedCrossValidationSubProblem.h
